refactor: search, explore 분리
